Selenium: Added setting to fail on SwitchScreen
The switchscreen function depends heavily on the browser settings for opening links in a new tab of screen.
Failure of this function is not always a failure of the test case. Added an option to choose whether to fail or not.
Name: prova.plugins.out.web.browser.failswitchscreen
Default setting = true.